updated kitchen tools table
changed table to have correct user id type
