+++ frontend/package.json
@@ -19,7 +19,7 @@
     "@react-navigation/bottom-tabs": "^7.3.10",
     "@react-navigation/elements": "^2.3.8",
     "@react-navigation/native": "^7.1.6",
-    "@stripe/stripe-react-native": "^0.56.0",
+    "@stripe/stripe-react-native": "0.50.3",
     "expo": "54.0.23",
     "expo-auth-session": "~7.0.8",
     "expo-blur": "~15.0.7",
